First container starts are broken
The output is different when containers have to be downloaded
and thus container IDs cannot be determined. This is a change
that came with the Docker upgrade to 1.3.

Also, Docker output should be forwarded to the user. Swallowing
the output results in a bad user experience because the user
is not informed about image downloads.

Fixes #6 and #8